settings: Don't try to disable flake8, document relevant limitation
flake8 plugin is disabled by default. If it's been enabled in the pylsp
configuration then it's not possible for another plugin to disable it.

The other four plugins (pycodestyle, pyflakes, mccabe, pyls_isort) are all
enabled by default so the ruff plugin can disable them, provided they're not
explicitly enabled in the configuration.

Closes #46