Run `ruff format` when lsp formatting is invoked (#57)
* Run `ruff format` when lsp formatting is invoked

Adds the Subcommand enum to indicate which `ruff` subcommand should be
executed by `run_ruff`. At this time, only `check` and `format` are
supported. As different subcommands support different parameters,
argument generation is delegated based on the specific subcommand value.

The `ruff format` subcommand does not currently organize imports and
there does not appear to be a way to convince it to do so. Until a
unified command exists the approach taken here is to format and then
make a second run of `ruff check` that _only_ performs import
formatting.

* Preserve compatibility with `format` settings

Codes listed in this setting should be included in fixes performed as
part of a formatting pass.

* Make import sorting opt-in